backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings>=2.0.0
email-validator
bcrypt==3.2.2
alembic
httpx>=0.27.0
itsdangerous>=2.1.0
psycopg2-binary>=2.9.9